# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Clean Architecture Example with Go, Docker.",
        "title": "Payment System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/businesses/delete/{id}": {
            "delete": {
                "description": "Soft delete a business (Logical Delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a Business",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Business UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/businesses/new": {
            "post": {
                "description": "Creates a new business entity with a specific commission rate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Register a new Business",
                "parameters": [
                    {
                        "description": "Business Registration",
                        "name": "business",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler.createBusinessRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Business"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/businesses/{id}": {
            "get": {
                "description": "Retrieve details of a specific business by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get Business Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Business UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Business"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Business not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/businesses/{id}/commission": {
            "patch": {
                "description": "Update the commission rate for an existing business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update Business Commission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Business UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Commission Rate",
                        "name": "commission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler.updateCommissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Business"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/audit": {
            "get": {
                "description": "Retrieve all system audit logs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Get All Logs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Log"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/audit/{resource_id}": {
            "get": {
                "description": "Retrieve audit logs for a specific resource (e.g., a Business ID or Transaction ID)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Get Audit Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Log"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_CardenalDex_crudprotec_internal_entitys.Business": {
            "type": "object",
            "properties": {
                "commission": {
                    "description": "Represented in basis points (e.g., 550 = 5.5%)",
                    "type": "integer",
                    "format": "int64"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_CardenalDex_crudprotec_internal_entitys.Log": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "\"UPDATE_TRANSACTION\"",
                    "type": "string"
                },
                "actor": {
                    "description": "made by",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "prevResourceID": {
                    "description": "Previous state",
                    "type": "string"
                },
                "resourceID": {
                    "description": "ID of the affected resource",
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_adapter_handler.createBusinessRequest": {
            "type": "object",
            "required": [
                "commission_percentage"
            ],
            "properties": {
                "commission_percentage": {
                    "description": "e.g., 5.5 for 5.5%",
                    "type": "number"
                }
            }
        },
...
        "internal_adapter_handler.updateCommissionRequest": {
            "type": "object",
            "required": [
                "new_commission_percentage"
            ],
            "properties": {
                "new_commission_percentage": {
                    "type": "number"
                }
            }
        }
    }
}